fix: scroll style
