+++ README.md
@@ -17,10 +17,10 @@
 - Иерархия слоёв FSD:
 /app (самый главный слой)
 └── /widgets  
-├── /pages
-├── /features
-├── /entities
-├── /shared
+├── /pages (каждая страница не должна содержать логику, а должна содержать фичи или виджеты — возможно из entities что-то используем)
+├── /features (бизнес логика)
+├── /entities (сущности)
+├── /shared (шарим функционал всем слоям)
 
 Слои:
 - app - самый главный, вышестоящий слой
