+++ src/shared/lib/hooks/index.ts
@@ -2,6 +2,6 @@ export { useFormRevalidateWithLocale } from './use-form-revalidate-with-locale'
 export { useLoader } from './use-loader'
 export { useResponsive } from './use-responsive'
 export { useTranslation } from './use-translation'
-export { useDebounce } from './use-debounce'
+export { debounce } from '@/shared/lib/utils/debounce'
 export { useHover } from './use-hover'
 export { useThrottle } from './use-throttle'
