+++ src/shared/lib/index.ts
@@ -7,4 +7,4 @@ export {
 } from '../../feature/translate'
 export { useFormRevalidateWithLocale, useLoader, useResponsive, useTranslation } from './hooks'
 export { HeadMeta, MetaNoIndex, onlyText } from './seo'
-export { cn, useDebounce, emailSchema, getIcon, mergeRefs, passwordSchema, throttle } from './utils'
+export { cn, debounce, emailSchema, getIcon, mergeRefs, passwordSchema, throttle } from './utils'
